Refactor: Create feature.Context
